requirements.txt
tflite-runtime==2.7.0
numpy==1.20.2
Pillow==8.1.0
requests==2.25.1
protobuf==3.14.0
streamlit==1.8.1
click==8.0.4
